finish lab02

+++ skeleton-fa22/lab02/DebugExercise/DebugExercise1.java
@@ -6,7 +6,7 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float) top / bottom;
         int result = Math.round(quotient);
         return result;
     }
